Fix SAdsNotificationHeader structure
The data field is a dynamically sized array that is part of the
structure. As such, declaring it as a pointer type is wrong.
If it is declared as a pointer, `bytearray(data)` is limited in size
to the size of a pointer on the platform Python is running on.
With x86 (32-bit) Python, it is not possible to capture double-
precision floating point values this way, since they are twice
as large as the pointer size. `struct.unpack` will then complain
that the sizes do not match up.
Instead, declare that the data field is one byte (like in the actual
C structure), and dynamically build a field of the right size
when it is known.
This additionally allows to remove the superfluous copy of
the value for string parsing.

This breaks backwards-compatibility when functions relied on
getting a pointer-sized byte array for unmapped types since
they are now also correctly sized.

Fixes #43