Fixes bug that causes an AmsPort in _adsDLL compiled library to be left open
On Linux systems, if an exception was thrown from the adsAddRoute(...) function call in the Connection.open() function, after a port has been opened, there was no call to close that port. This lack of closing the port renders the underlying AmsPort to be unusable for later connections, and if all 128 AmsPorts end up in this state, all Connection objects are rendered useless.